Add clasification worker + display result

+++ serialize.py
@@ -1,7 +1,9 @@
 import json
+import torch
+import numpy as np
 
 def torch_to_json(tensor):
     return json.dumps(tensor.numpy().tolist())
 
-def json_to_torch(data):
-    return torch.tensor(np.array(json.loads(data)))
+def json_to_torch(data, dtype=torch.float):
+    return torch.tensor(np.array(json.loads(data)), dtype=dtype)
